refactor: refactor code and file-name
- update project dependencies
- now the service files include `.service` ext on filename
- change filename from camelCase to kebabCase
- remove socket code
- remove string.js file
- move StateService to `storage` directory
- move DATABASE_COLUMNS to StorageKey file
