+++ package.json
@@ -97,19 +97,17 @@
     "test": "node ./out/test/runTest.js"
   },
   "devDependencies": {
-    "@types/glob": "^7.1.1",
-    "@types/mocha": "^5.2.7",
-    "@types/node": "^12.11.7",
-    "@types/socket.io-client": "^1.4.32",
-    "@types/vscode": "^1.40.0",
-    "glob": "^7.1.5",
-    "mocha": "^6.2.2",
-    "tslint": "^5.20.0",
-    "typescript": "^3.6.4",
-    "vscode-test": "^1.2.2"
+    "@types/glob": "^7.1.4",
+    "@types/mocha": "^9.0.0",
+    "@types/node": "^16.7.2",
+    "@types/vscode": "^1.59.0",
+    "glob": "^7.1.7",
+    "mocha": "^9.1.0",
+    "typescript": "^4.3.5",
+    "tslint": "^6.1.3",
+    "vscode-test": "^1.6.1"
   },
   "dependencies": {
-    "axios": "^0.19.0",
-    "socket.io-client": "^2.3.0"
+    "axios": "^0.21.1"
   }
 }
