+++ src/services/storage/storage-key.ts
@@ -1,4 +1,4 @@
-export const DATABASE_COLUMNS = {
+export const StorageKey = {
   activeUserId: 'activeUserId',
   sailsSessionId: 'sails.sid'
 }
